chore(generator): added more tests

+++ openapi-generator/lib/src/openapi_generator_runner.dart
@@ -93,25 +93,28 @@ class OpenapiGenerator extends GeneratorForAnnotation<annots.Openapi> {
       arguments.cleanSubOutputDirectory?.forEach((directory) {
         final childDirectory =
             Directory('${arguments.outputDirectory}/${directory.toString()}');
-        try {
-          final outputDirectory = Directory('${arguments.outputDirectory}')
-              .resolveSymbolicLinksSync();
-          // Make sure is sub directory, avoid ../../
-          if (childDirectory
-              .resolveSymbolicLinksSync()
-              .startsWith(outputDirectory)) {
-            childDirectory.delete(recursive: true);
+        if (childDirectory.existsSync() &&
+            childDirectory.listSync().isNotEmpty) {
+          try {
+            final outputDirectory = Directory('${arguments.outputDirectory}')
+                .resolveSymbolicLinksSync();
+            // Make sure is sub directory, avoid ../../
+            if (childDirectory
+                .resolveSymbolicLinksSync()
+                .startsWith(outputDirectory)) {
+              childDirectory.delete(recursive: true);
+            }
+          } catch (e, st) {
+            logOutputMessage(
+              log: log,
+              communication: OutputMessage(
+                message: 'Output directory already empty',
+                additionalContext: e,
+                stackTrace: st,
+                level: Level.WARNING,
+              ),
+            );
           }
-        } catch (e, st) {
-          logOutputMessage(
-            log: log,
-            communication: OutputMessage(
-              message: 'Output directory already empty',
-              additionalContext: e,
-              stackTrace: st,
-              level: Level.WARNING,
-            ),
-          );
         }
       });
     }
